requirements.txt
flask==2.3.3
flask-cors==4.0.0
influxdb-client==1.38.0
pandas==2.1.1
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.5
schedule==1.2.0
geoip2==4.7.0

